+++ iot/iot-device-registry-infinispan/src/main/java/io/enmasse/iot/registry/infinispan/InfinispanRegistry.java
@@ -20,10 +20,9 @@ import org.springframework.context.annotation.ComponentScan;
 @ComponentScan("org.eclipse.hono.service.auth")
 @ComponentScan("org.eclipse.hono.service.metric")
 @ComponentScan("io.enmasse.iot.registry.infinispan")
+@ComponentScan("io.enmasse.iot.service.base")
 @EnableAutoConfiguration
 public class InfinispanRegistry extends AbstractBaseApplication {
-
-    public static final String CONFIG_BASE = "enmasse.iot";
 
     /**
      * All the verticles.
