+++ iot/iot-device-registry-infinispan/src/main/java/io/enmasse/iot/registry/infinispan/debug/DebugEndpoint.java
@@ -15,7 +15,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
-import io.enmasse.iot.registry.infinispan.InfinispanRegistry;
+import io.enmasse.iot.service.base.infinispan.config.InfinispanProperties;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
@@ -26,7 +26,7 @@ import io.vertx.ext.web.RoutingContext;
 
 @Component
 @ConditionalOnProperty("debug")
-@ConfigurationProperties(InfinispanRegistry.CONFIG_BASE + ".registry.debug")
+@ConfigurationProperties(InfinispanProperties.CONFIG_BASE + ".registry.debug")
 public class DebugEndpoint extends AbstractVerticle {
 
     @Autowired
