+++ iot/iot-device-registry-infinispan/src/main/java/io/enmasse/iot/registry/infinispan/tenant/TenantInformationService.java
@@ -5,6 +5,7 @@
 
 package io.enmasse.iot.registry.infinispan.tenant;
 
+import io.enmasse.iot.service.base.infinispan.tenant.TenantHandle;
 import java.util.concurrent.CompletableFuture;
 
 import io.opentracing.Span;
