+++ iot/iot-service-base/src/main/java/io/enmasse/iot/service/base/infinispan/devcon/DeviceConnectionKey.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.iot.registry.infinispan.devcon;
+package io.enmasse.iot.service.base.infinispan.devcon;
 
 import java.io.Serializable;
 import java.util.Objects;
